backend/requirements.txt
fastapi
uvicorn
python-multipart
nltk
huggingface-hub
python-dotenv